Update index.html
updated 5 functionality, 1. make the contact form responsive in the response i can the sense that I can receive the message sent via the form 2. Get in touch section, add below something like we are a team experts and can help with your model distillation needs, use the right language 3. are there more content you can think of for key resources section, say like 5-7 compared to currently 3 listed 4. if we add 5-7 or more resource, would it be better to add some scroll functionality? do the best in terms of user experience and engagement 5. it would be good add another section on "Case Studies" where we can highlight 3-5 most compelling successful widely popular examples of model distillation, find these resources and add and the whole section to the website if it is a good idea